order 재고 관리 추가 및 코드 정리

+++ src/main/java/spring/webmvc/application/dto/command/ProductCreateCommand.java
@@ -10,6 +10,6 @@ public class ProductCreateCommand {
 	private final Category category;
 	private final String name;
 	private final String description;
-	private final int price;
-	private final int quantity;
+	private final long price;
+	private final long quantity;
 }
