+++ src/main/java/spring/webmvc/application/dto/result/ProductResult.java
@@ -14,8 +14,8 @@ public class ProductResult {
 	private final Category category;
 	private final String name;
 	private final String description;
-	private final int price;
-	private final int quantity;
+	private final long price;
+	private final long quantity;
 	private final Instant createdAt;
 
 	public ProductResult(Product product) {
